add custom hook for fetch. fix api fetch bug
