+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
 import About from './Components/Nav/About';
@@ -6,31 +6,13 @@ import Nav from './Components/Nav/Nav';
 import Main from './Components/Main/Main';
 import RecipeDetails from './Components/RecipeDetails/RecipeDetails';
 
-import randomRecipes1 from './data.json';
 
 function App() {
-	const [randomRecipes, setRandomRecipes] = useState([]);
-
-	useEffect(() => {
-	    fetchRandomRecipes(3);
-	}, [])
-
-	const fetchRandomRecipes = (num) => {
-		let url = `https://api.spoonacular.com/recipes/random?number=${num}&apiKey=${process.env.REACT_APP_API_KEY}`;
-		fetch(url)
-			.then((res) => res.json())
-			.then((res) => {
-				console.log(res.recipes);
-				setRandomRecipes(res.recipes);
-			});
-	};
-
 	return (
 		<div>
 			<Nav />
-
 			<Routes>
-        <Route path='/' element={<Main randomRecipes={randomRecipes} />} />
+        <Route path='/' element={<Main />} />
 				<Route path='/:id' element={<RecipeDetails />} />
 				<Route path='/about' element={<About />} />
 			</Routes>
