+++ src/Components/RecipeDetails/RecipeDetails.css
@@ -1,0 +1,3 @@
+body {
+    background-color: #E3E3E3;
+}
