+++ src/index.css
@@ -1,11 +1,8 @@
-
-
+a {
+  text-decoration: none;
+}
 body {
   background-image: url(./Components/nordwood-themes-Tmz8FThN_BE-unsplash.jpg);
   background-size: cover;
   /* text-align: center; */
-}
-
-a {
-  text-decoration: none;
 }
